feat: regenerate database upgrade scripts
