+++ src/alembic/versions/cb1df1f5c488_string_revision_ids.py
@@ -1,0 +1,62 @@
+"""string revision ids
+
+Revision ID: cb1df1f5c488
+Revises: ddacf7741794
+Create Date: 2026-05-26 12:38:40.654579
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = 'cb1df1f5c488'
+down_revision: Union[str, Sequence[str], None] = 'ddacf7741794'
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    """Upgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    with op.batch_alter_table('document_revisions', schema=None) as batch_op:
+        batch_op.alter_column('id',
+               existing_type=sa.INTEGER(),
+               type_=sa.VARCHAR(length=64),
+               existing_nullable=False)
+        batch_op.alter_column('parent_revision_id',
+               existing_type=sa.INTEGER(),
+               type_=sa.VARCHAR(length=64),
+               existing_nullable=True)
+
+    with op.batch_alter_table('documents', schema=None) as batch_op:
+        batch_op.alter_column('current_revision_id',
+               existing_type=sa.INTEGER(),
+               type_=sa.VARCHAR(length=64),
+               existing_nullable=True)
+
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    """Downgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    with op.batch_alter_table('documents', schema=None) as batch_op:
+        batch_op.alter_column('current_revision_id',
+               existing_type=sa.VARCHAR(length=64),
+               type_=sa.INTEGER(),
+               existing_nullable=True)
+
+    with op.batch_alter_table('document_revisions', schema=None) as batch_op:
+        batch_op.alter_column('parent_revision_id',
+               existing_type=sa.VARCHAR(length=64),
+               type_=sa.INTEGER(),
+               existing_nullable=True)
+        batch_op.alter_column('id',
+               existing_type=sa.VARCHAR(length=64),
+               type_=sa.INTEGER(),
+               existing_nullable=False)
+
+    # ### end Alembic commands ###
